# backend/openapi.yaml
openapi: 3.1.0
info:
  title: UniversalCart API
  version: 0.5.0
  description: |
    Universal shopping cart Chrome extension backend API.
    Provides cart management, price scouting, checkout orchestration,
    promo code aggregation, deal finding, notifications, and security services.
  contact:
    name: UniversalCart
    url: https://github.com/shroominati/universalcart
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.universalcart.app
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from /api/auth/login or /api/auth/register
    RequestSigning:
      type: apiKey
      in: header
      name: X-UC-Signature
      description: HMAC-SHA256 request signature for extension ↔ backend calls
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        code:
          type: string
          example: VALIDATION_ERROR
        details:
          type: array
          items:
            type: object
            properties:
              location:
                type: string
                enum: [body, params, query]
              field:
                type: string
              message:
                type: string
      required: [error, code, details]
    RateLimitError:
      type: object
      properties:
        error:
          type: string
        retryAfter:
          type: integer
          description: Seconds until rate limit resets
...
    PaymentMethod:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [card, paypal, applepay, googlepay]
          default: card
        last4:
          type: string
          pattern: "^\\d{4}$"
        brand:
          type: string
        expiryMonth:
          type: integer
          minimum: 1
          maximum: 12
        expiryYear:
          type: integer
          minimum: 2024
          maximum: 2040
        isDefault:
          type: boolean
          default: false
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [price_drop, deal_found, back_in_stock, checkout, system]
        title:
          type: string
        message:
          type: string
        data:
          type: object
        priority:
          type: string
          enum: [low, medium, high]
        read:
          type: boolean
        createdAt:
          type: integer
    AuditEntry:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: integer
        type:
          type: string
        userId:
          type: string
        action:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        chainHash:
          type: string
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    VendorParam:
      name: vendor
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - BearerAuth: []
paths:
...
  /api/checkout/start:
    post:
      tags: [Checkout]
      summary: Start automated multi-vendor checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cart:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      price:
                        type: number
                      sourceUrl:
                        type: string
                        format: uri
                      vendor:
                        type: string
                      quantity:
                        type: integer
                        default: 1
                    required: [name, sourceUrl, vendor]
                  minItems: 1
                shippingProfileId:
                  type: string
                paymentMethodId:
                  type: string
                applyPromos:
                  type: boolean
                  default: true
                dryRun:
                  type: boolean
                  default: false
              required: [cart, shippingProfileId, paymentMethodId]
      responses:
        "202":
          description: Checkout job started
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                  status:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/checkout/jobs/{id}:
    get:
      tags: [Checkout]
      summary: Get checkout job status
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Job status
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Checkout]
      summary: Cancel checkout job
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Job cancelled
...
  /api/checkout/profiles/payment:
    post:
      tags: [Checkout]
      summary: Create payment method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethod"
      responses:
        "201":
          description: Payment method created
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/checkout/profiles/payment/{id}:
    delete:
      tags: [Checkout]
      summary: Delete payment method
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Payment method deleted

  # ─── Promos ──────────────────────────────────────────────────────────────────
  /api/promos/{vendor}:
    get:
      tags: [Promos]
      summary: Get promo codes for a vendor
      parameters:
        - $ref: "#/components/parameters/VendorParam"
      responses:
        "200":
          description: Promo codes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code:
                      type: string
                    description:
                      type: string
                    source:
                      type: string
                    verified:
                      type: boolean
    post:
      tags: [Promos]
      summary: Submit a promo code
      parameters:
        - $ref: "#/components/parameters/VendorParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
                source:
                  type: string
                  maxLength: 100
              required: [code]
      responses:
        "201":
          description: Promo submitted
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/promos/{vendor}/scan:
    post:
      tags: [Promos]
      summary: Trigger promo code scan for vendor
      parameters:
        - $ref: "#/components/parameters/VendorParam"
      responses:
        "202":
          description: Scan started

  # ─── Price Scout ─────────────────────────────────────────────────────────────
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Get user notifications
      parameters:
        - name: unread
          in: query
          schema:
            type: string
            enum: ["true", "false"]
        - name: type
          in: query
          schema:
            type: string
            enum: [price_drop, deal_found, back_in_stock, checkout, system]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /api/notifications/{id}/read:
    post:
      tags: [Notifications]
      summary: Mark notification as read
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Marked as read
  /api/notifications/{id}:
    delete:
      tags: [Notifications]
      summary: Delete notification
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
  /api/notifications/preferences:
    patch:
      tags: [Notifications]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                priceDrop:
                  type: boolean
                dealFound:
                  type: boolean
                backInStock:
                  type: boolean
                checkout:
                  type: boolean
                emailDigest:
                  type: boolean
                digestFrequency:
                  type: string
                  enum: [daily, weekly, never]
      responses:
        "200":
          description: Preferences updated
        "400":
          $ref: "#/components/responses/BadRequest"

  # ─── Security ────────────────────────────────────────────────────────────────
  /api/security/audit:
    get:
      tags: [Security]
      summary: Query audit log
      parameters:
        - name: type
          in: query
          schema:
            type: string
            maxLength: 50
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, critical]
        - name: since
          in: query
          schema:
            type: integer
            description: Unix timestamp ms
        - name: until
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
  /api/security/keys/{service}/rotate:
    post:
      tags: [Security]
      summary: Rotate API key for a service
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                newKey:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Key rotated

  # ─── Analytics ───────────────────────────────────────────────────────────────
...